+++ public/css/login.module.css
@@ -33,6 +33,14 @@
     z-index: -1;
 }
 
+.formWrapper {
+    width: 75%;
+    height: 75%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
 .form {
     display: flex;
     flex-direction: column;
@@ -43,7 +51,7 @@
     box-shadow: 0 4px 30px rgba(0, 0, 0, 1);
     backdrop-filter: blur(14.8px);
     -webkit-backdrop-filter: blur(14.8px);
-    height: 60%;
+    height: 80%;
     width: 35%;
 }
 
@@ -80,7 +88,7 @@
     width: 100%;
 
     &::placeholder {
-        color: #fff;
+        color: rgba(255, 255, 255, 0.3);
     }
 
     &:focus {
@@ -125,12 +133,26 @@
     }
 }
 
-@media only screen and (max-width: 1000px) {
-    .form {
+@media only screen and (max-width: 1200px) {
+
+    .formWrapper {
         width: 100%;
         height: 100%;
-        margin: 0;
+        align-items: flex-end;
+        background: rgba(0, 0, 0, 0.3);
         border-radius: 0;
+        box-shadow: 0 4px 20px rgba(0, 0, 0, 1);
+        backdrop-filter: blur(15px);
+    }
+
+    .form {
+        background: none;
+        box-shadow: none;
+        backdrop-filter: none;
+        width: 100%;
+        height: 75%;
+        border-radius: 0;
+        justify-content: space-evenly;
     }
 
     .title {
@@ -140,8 +162,6 @@
     .inputs {
         justify-content: center;
         gap: 10px;
-        position: absolute;
-        bottom: 290px;
         width: 75%;
     }
 
@@ -157,8 +177,6 @@
 
     .buttons {
         flex-direction: column;
-        position: absolute;
-        bottom: 100px;
         gap: 10px;
         font-size: 14px;
         margin: 0;
